requirements.txt
python-telegram-bot
aiohttp
asyncio

